+++ app/src/main/java/org/takingroot/assignment/Constants.kt
@@ -1,0 +1,7 @@
+package org.takingroot.assignment
+
+object Constants {
+
+    const val accountIdToken = "344f04dc-11f5-4b40-aaa0-06719d02dc03"
+    const val baseUrl = "http://assignment.takingroot.app"
+}
